Refactored getChunks and notepadToHTML to pad-utils
